+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Change type of service keys from `string` to `interface{}`. [#4](https://github.com/go-nacelle/service/pull/4)
+
 ### Removed
 
 - Removed MustGet and MustSet methods from the Container interface. [#3](https://github.com/go-nacelle/service/pull/3)